+++ FitnessTracker.Core/Dtos/Requests/Users/EditUserRequest.cs
@@ -10,8 +10,8 @@ public record EditUserRequest
     public string? Email { get; init; }
     public string? Phone { get; init; }
     public DateOnly? Birthday { get; init; }
-    public double? Height { get; init; }
-    public double? Weight { get; init; }
+    public decimal? Height { get; init; }
+    public decimal? Weight { get; init; }
     public Gender? Gender { get; init; }
 }
