+++ FitnessTracker.Core/Dtos/Responses/Users/ShortUserResponse.cs
@@ -7,7 +7,7 @@ public record ShortUserResponse
     public Guid Id { get; init; }
     public required string Name { get; init; }
     public DateOnly Birthday { get; init; }
-    public double Height { get; init; }
-    public double Weight { get; init; }
+    public decimal Height { get; init; }
+    public decimal Weight { get; init; }
     public Gender Gender { get; init; }
 }
